Move encoder class into utils
